+++ common/src/fonts.h
@@ -10,7 +10,9 @@
   // Include custom font files for larger displays
   #include <fonts/FreeSans9pt7b.h>
   #include <fonts/FreeSans12pt7b.h>
+  #include <fonts/FreeSans18pt7b.h>
   #include <fonts/FreeSansBold18pt7b.h>
+  #include <fonts/FreeSansBold24pt7b.h>
 #endif
 
 #endif // FONTS_H